requirements.txt
mysql-connector-python
tqdm
